django-assignment/requirements.txt
asgiref==3.5.1
Django==3.2.13
django-cors-headers==3.12.0
djangorestframework==3.13.1
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.2.0
